refactoring update checker
